requirements.txt
click==8.1.3
colorama==0.4.6
Pillow==9.5.0
PyYAML==6.0
simple-threadpool==1.0.0
